one error left in rust program anaheim

+++ anchor/programs/anaheim/src/state/anaheim_account.rs
@@ -3,7 +3,7 @@ use anchor_lang::prelude::*;
 #[account]
 pub struct Anaheim {
   pub authority: Pubkey,
-  pub value: u8,
+  pub value: u64,
   pub count: u64,
   pub counter: u8, // <-- il te faut CE champ si tu veux le manipuler
 }
